Add new ldap_network_timeout config option
Introduction of PHP 5.3 support with 1.3.x allows setting of LDAP
network timeout (LDAP_OPT_NETWORK_TIMEOUT) before binding the server.

When the LDAP server is not available with this option set to its
default value of 0 (infinite), the system will seem to hang for several
minutes, depending on the TCP stack settings, until control is finally
returned to Mantis.

Setting this option to a low value makes the system more responsive,
and also allows automatic and rapid failover to the next available LDAP
server when the hostname defined in $g_ldap_server resolves to multiple
IP addresses, which are tried in sequence by ldap_bind().

Fixes #12544