Enable menu 'Report Issue' if user can report in any project
Prior to this, the 'Report Issue' menu option was not available if
current project was 'All Projects' and the user's global access level
did not allow them to report issues.

This commit changes the logic to show the menu, and displays it whenever
the user has report_bug_threshold in any of the projects they have
access to.

This is achieved via new function access_can_report_bug() in access API.
To avoid potential performance issues in installation having a large
number of projects, the results are cached.

Fixes #14268