Browser search plug-in branding
The title used in the browser search plug-in can now be customized. By
default it's still MantisBT. The rest of the strings used in the plug-in
can be localized.

The favicon used by the plugin will be the favicon configured in
MantisBT. If not present, the default icon has been updated to the
current MantisBT project icon.